MMC/SD: SDIO reset functionality to the existing MMC frame work
Observed some issues during  the initialization of SDIO cards. The
initialization sequence for SDIO card is not according to the SDIO
specification. Hence the card is not getting initialized during module
reload with card always present in the slot. This patch fixes the
issue.

Signed-off-by: phani@embwise.com

This patch also changes the dev_err to dev_dbg for the CMD timeout error
messages. This was changed earlier in one of the SDIO patches.

Signed-off-by: Sneha Narnakaje <nsnehaprabha@ti.com>