davinci: dm646x: Add platform data for NAND
This patch adds platform data and partition info for NAND on dm6467 EVM.

Note that the partition layout is dependent on the UBL, U-Boot combination used.
This patch uses partition organization suitable with latest U-Boot of LSP 1.3.
For example, U-Boot environment goes in block 0, UBL resides in block form 1 to
5 and so on.

Signed-off-by: Hemant Pedanekar <hemantp@ti.com>