V4L: vpfe_capture - bug fixes
Fixes following:-
	1) free contig memory only for MMAP
	2) use videobuf_iolock() to get physical memory for user page

Signed-off-by: Muralidharan Karicheri <m-karicheri2@ti.com>

The patch also fixes the crash issue with buffer allocation, which
happened due to not passing the device structure to the
dma_contig_init call. It is normal to have buffer allocation
failures for HD resolutions, but the VPFE capture driver should
gracefully exit and not crash.

Signed-off-by: Sneha Narnakaje <nsnehaprabha@ti.com>

The patch also fixes the issue with vpfe_videobuf_release handler not to
flush the vpfe capture buffer queue. Flushing the queue caused the loopback
application to hang after 3 frames.

Signed-off-by: Sandeep Paulraj <s-paulraj@ti.com>