mtd-nand: davinci: Correct 4-bit error correction
On TI's DA830/OMAP-L137 and DA850/OMAP-L138, after setting
the 4BITECC_ADD_CALC_START bit in the NAND Flash control
register to 1 and before waiting for the NAND Flash status
register to be equal to 1, 2 or 3, we have to wait till the
ECC HW goes to correction state. Without this wait, ECC
correction calculations will not be proper.

This has been tested on DA830/OMAP-L137 and DA850/OMAP-L138
EVMs.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>