V4L - Support for non-standard resolution capture and display
When a resolution of 176x144 is used in S_FMT, the dut crash and never recovers
This was because, the line length and size was not set correctly in ccdc. This
results in hardware overwriting outside the buffer address space resulting in
the crash. Also display side the videobuf_iolock() fails because the vb sizes
are not correct. Both these issues are fixed in this patch so that application
can do non-standard resolution capture and display.

Test: Use S_FMT with above resolution for capture and display in the v4l2 device.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>