davinci: edma: clear interrupt status for interrupt enabled channels only
Currently, the ISR in the EDMA driver clears the pending interrupt
for all channels without regard to whether that channel has a registered
callback or not.

This causes problems when DSP side codecs who really do not require
an ARM side callback poll on the interrupt pending bits of some of
the channels used by them. There is an obvious race between the ISR
on the ARM side and the codecs on the DSP side.

This patch fixes the issue by making the ISR clear the interrupts only
for those channels which have interrupt enabled in shadow region 0.

Signed-off-by: Anuj Aggarwal <anuj.aggarwal@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>