davinci: dm6467t: change default Luma channel in THS7353 to 2
The Rev C of DM6467T EVM changed the ordering of connections
in THS7353 to make it compatible with DM6467 EVM.

This patch updates the defconfig to work correctly on the
Rev C board. By definition this *breaks* compatibility with
Rev A and B of DM6467T EVM. Users of those boards should
set the Luma channel configuration to 3.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>