asoc: davinci: let platform data specify edma queue number
Currently the queue number to be used by EDMA channel
servicing audio data is fixed to EDMAQ_0. This is not
desirable on some platforms like DM365 where Queue 0
has large fifo size suitable for video transfers.

The platform data for audio does define the queue number
but that remains unused inside the audio driver.

If no platform data is passed, EDMAQ_0 is used.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>