davinci: add support for aemif timing configuration
This patch adds support to configure the AEMIF interface
for user supplied timing values.

Since this capability is useful both from NOR and NAND
flashes, it is provided as a new interface in a file of
its own.

Earlier, AEMIF configuration was being done as a special case
for DM644x board, but now more boards emerge which have capability
to boot for other media (SPI flash, NOR flash) and have the kernel
access NAND flash. This means that kernel cannot always  depend on
the bootloader to setup the NAND.

Also, on platforms such as da850/omap-l138, the aemif input
frequency changes as cpu frequency changes; necessiating
re-calculation of timimg values as part of cpufreq transtitions.
This patch forms the basis for adding that support.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>