dm365: Remove MMC/SD card detect function
On DM365 CPLD is being used to detect whether MMC/SD card is
present in the slot or not. CPLD has been interfaced using EMIF
(like NAND). The MMC/SD card detect function gets called at
every timer interrupt interval. Because of this frequent call,
nandtest fails.

To get around this problem, this patch disables the card detect
function in arch/arm/mach-davinci/board-dm365-evm.c file.This
will not affect the MMC/SD functionality on the EVM, as there
is a card detect function in the MMC framework which detects
whether the card is present or not.

This change is being done as a temporary workaround while
the real solution to sharing the AEMIF is found. Note that
the get_ro() is interfaced through CPLD as well, but that
function doesnt create a problem of the same magnitude
since it is called only once after a card is detected.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>