Replace usage of cbcr_ofst with davinci_fb_desc.
The value passed to VIDIOC_S_COFST is currently stored in a cbcr_ofst value in
the display manager, and is passed directly to lower OSD levels by adding a
cbcr_ofst parameter where needed.

This change generalizes that infrastructure by creating a new davinci_fb_desc
structure to contain frame buffer information such as cbcr_ofst, and the
existing infrastructure that passes information the the lower OSD levels was
updated to pass this instead.

This allows re-use of this infrastructure by new ioctls to communicate
information about the frame buffer with minimal modifications.