dm365: Add custom display ioctl VIDIOC_S_YDOFST
This ioctl specifies the offset into the Y-plane where the display should
start.  This is useful when displaying codec output buffers that have
padding around the edges of the frame.

The offset must be 64-byte aligned.  The hardware requirement is that both
the Y and C planes are 32-byte aligned.  However, the Y plane offset must be
64-byte aligned in order to guarantee that the corresponding C plane offset is
32-byte aligned.