dm368: Configure EXTCLK as clock source for HD video clock
On DM368, ARM/DDR should operate at 432MHz/340MHz respectively.
The result of configuring the PLLs for those frequencies is
that there is no clock that divides evenly to get a 74.25MHz
HD video clock. So on the DM368, we use the 74.25MHz clock
(from CPLD) going into EXTCLK and use that as the VPBE clock
source.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>