rtc: pcf8563: return EINVAL if time not valid
If the hardware clock contains a not reliable date or time it is
not correct to syncronize the system clock to it.
With this patch in case of not reliable time in the hardware clock
the system will start using 'Thu Jan  1 00:00:00 UTC 1970' date.
The system clock time is good to allow the system working.
The application, hwclock or similar, is able, using -EINVAL return
value, to recognize that the system clock is not good and can
syncronize using alternative ways (ntp, ...).
After it it will syncronize back the hardware clock using
the good system clock.

Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>