spi: DaVinci: fixed bug using more devices on the same peripheral
Signed-off-by: Davide Bonfanti <davide.bonfanti@bticino.it>
Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>

----
When transmitting using a chip-select the other was disabled (clearing
SPIPC0.SCSFUN and SPIPC0.SCS1FUN). In this way the chip-select that was not
used was tied to ground when the peripheral was activated causing a the
chip-select enable (usually it is ative low!).
Now SPIPC0.SCSFUN and SPIPC0.SCS1FUN are always enabled.