DaVinci: Basi Board: pull down disabled
Pull-downs can give problems, so I disable all of them.

For example the pull-down is the default state for CIN7 DM365 port and it gives
a voltage drop of 1.1V on I2CSEL TVP5151 input causing sometime the erroneous
setting of I2C address.
With this patch the pull-down is disabled and I2CSEL is at its correct value.

Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>