Or'ed gpio interrupt: initial support
When the cpu offers less interrupt-capable gpio than what needed by the board,
they can be or'ed together. In this way a falling edge on the or'ed pin
generates an hw interrupt, in the interrupt service routine all the gpio are
controlled and the proper interrupt is sent to the system.

Signed-off-by: Davide Bonfanti <davide.bonfanti@bticino.it>
Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>