DaVinci: dm365: added sdram dma clock divide ratio management
The CLKDIV register is then used to select a divide ratio of the SDRAM(DMA)
clock for the pixel clock frequency which is used to clock the data into the
PCLK.
The CLKDIV register is then used to select a divide ratio of the SDRAM(DMA)
clock for the pixel clock frequency which is used to clock the data into the
PCLK. The value of this register depends on the resize ratios of the IPIPE
resizers and the available SDRAM system bandwidth.
Once defined input and output resolutions to the Resizer, the clock must be
at least MAX(input_resolution, output_resolution) * framerate. This is a rough
approximation, the resizer needs some additional lines and pixels at the frames
edges so a slightly higher clock than above is needed. A 12.5% of increment is
implemented.
If resizer clock is too high, then what will happen is that resizer will try to
read pixels at every clock cycle. If there is someone else running in the
system (VPSS capture or ARM or codec) then if the resizer will fail to read the
pixel at any given moment due to other master using the DDR at that time, IT
WILL ABORT the operation causing an incomplete frame in memory.

Signed-off-by: Davide Bonfanti <davide.bonfanti@bticino.it>