irq: irq_gpio: toggle between fall and raise gpio edge using a variable
If there is some glitches on the gpio it can happen that when a falling/rising
interrupt occurs when the irq code is executed a gpio sampling can give the
wrong state. For this reason it is assumed that at start-up the interrupt gpio
is high and then its state is supposed to be toggled at each interrupt
occurrence. Before this patch the edge that caused the interrupt was inferred
reading the gpio state in the isr routine.

Signed-off-by: Davide Bonfanti <davide.bonfanti@bticino.it>
Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>