V4L/DVB: V4L2: add a new V4L2_CID_EXPOSURE_WINDOWS_WEIGHT menu control
Add a new V4L2_CID_EXPOSURE_WINDOWS_WEIGHT menu control, which can be used
to set the weight of sections in a sensor while calculating auto-exposure
time starting from brigthness average value. Usually the frame is divided
in a grid of windows whose brightness can be differently weighted in order
to calculate the sensor auto-exposure.

Signed-off-by: Davide Bonfanti <davide.bonfanti@bticino.it>
Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>